requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.10
flask-cors==6.0.1
python-dotenv==1.1.1
gunicorn==23.0.0
Werkzeug==3.1.3
